tinystr

Struct TinyAsciiStr

pub struct TinyAsciiStr<const N: usize> { /* private fields */ }

Implementations

impl<const N: usize> TinyAsciiStr<N>

pub const fn from_bytes(bytes: &[u8]) -> Result<Self, TinyStrError>

Creates a TinyAsciiStr<N> from the given byte slice. bytes may contain at most N non-null ASCII bytes.

pub const fn from_bytes_lossy(bytes: &[u8]) -> Self

Creates a TinyAsciiStr<N> from a byte slice, replacing invalid bytes.

Null and non-ASCII bytes (i.e. those outside the range 0x01..=0x7F) will be replaced with the ‘?’ character.

The input slice will be truncated if its length exceeds N.

pub const fn try_from_raw(raw: [u8; N]) -> Result<Self, TinyStrError>

Attempts to parse a fixed-length byte array to a TinyAsciiStr.

The byte array may contain trailing NUL bytes.

Example

use tinystr::tinystr;
use tinystr::TinyAsciiStr;

assert_eq!(
    TinyAsciiStr::<3>::try_from_raw(*b"GB\0"),
    Ok(tinystr!(3, "GB"))
);
assert_eq!(
    TinyAsciiStr::<3>::try_from_raw(*b"USD"),
    Ok(tinystr!(3, "USD"))
);
assert!(matches!(TinyAsciiStr::<3>::try_from_raw(*b"\0A\0"), Err(_)));

pub const fn from_bytes_manual_slice( bytes: &[u8], start: usize, end: usize, ) -> Result<Self, TinyStrError>

Equivalent to from_bytes(bytes[start..end]), but callable in a const context (which range indexing is not).

pub const fn from_str(s: &str) -> Result<Self, TinyStrError>

pub const fn as_str(&self) -> &str

pub const fn len(&self) -> usize

pub const fn is_empty(&self) -> bool

pub const fn as_bytes(&self) -> &[u8]

pub const fn all_bytes(&self) -> &[u8; N]

pub const fn resize<const M: usize>(self) -> TinyAsciiStr<M>

Resizes a TinyAsciiStr<N> to a TinyAsciiStr<M>.

If M < len() the string gets truncated, otherwise only the memory representation changes.

pub const unsafe fn from_bytes_unchecked(bytes: [u8; N]) -> Self

Safety

Must be called with a bytes array made of valid ASCII bytes, with no null bytes between ASCII characters

impl<const N: usize> TinyAsciiStr<N>

pub const fn is_ascii_alphabetic(&self) -> bool

Checks if the value is composed of ASCII alphabetic characters:

U+0041 ‘A’ ..= U+005A ‘Z’, or

U+0061 ‘a’ ..= U+007A ‘z’.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "Test".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "Te3t".parse().expect("Failed to parse.");

assert!(s1.is_ascii_alphabetic());
assert!(!s2.is_ascii_alphabetic());

pub const fn is_ascii_alphanumeric(&self) -> bool

Checks if the value is composed of ASCII alphanumeric characters:

U+0041 ‘A’ ..= U+005A ‘Z’, or

U+0061 ‘a’ ..= U+007A ‘z’, or

U+0030 ‘0’ ..= U+0039 ‘9’.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "A15b".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "[3@w".parse().expect("Failed to parse.");

assert!(s1.is_ascii_alphanumeric());
assert!(!s2.is_ascii_alphanumeric());

pub const fn is_ascii_numeric(&self) -> bool

Checks if the value is composed of ASCII decimal digits:

U+0030 ‘0’ ..= U+0039 ‘9’.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "312".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "3d".parse().expect("Failed to parse.");

assert!(s1.is_ascii_numeric());
assert!(!s2.is_ascii_numeric());

pub const fn is_ascii_lowercase(&self) -> bool

Checks if the value is in ASCII lower case.

All letter characters are checked for case. Non-letter characters are ignored.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "teSt".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "test".parse().expect("Failed to parse.");
let s3: TinyAsciiStr<4> = "001z".parse().expect("Failed to parse.");

assert!(!s1.is_ascii_lowercase());
assert!(s2.is_ascii_lowercase());
assert!(s3.is_ascii_lowercase());

pub const fn is_ascii_titlecase(&self) -> bool

Checks if the value is in ASCII title case.

This verifies that the first character is ASCII uppercase and all others ASCII lowercase. Non-letter characters are ignored.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "teSt".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "Test".parse().expect("Failed to parse.");
let s3: TinyAsciiStr<4> = "001z".parse().expect("Failed to parse.");

assert!(!s1.is_ascii_titlecase());
assert!(s2.is_ascii_titlecase());
assert!(s3.is_ascii_titlecase());

pub const fn is_ascii_uppercase(&self) -> bool

Checks if the value is in ASCII upper case.

All letter characters are checked for case. Non-letter characters are ignored.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "teSt".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "TEST".parse().expect("Failed to parse.");
let s3: TinyAsciiStr<4> = "001z".parse().expect("Failed to parse.");

assert!(!s1.is_ascii_uppercase());
assert!(s2.is_ascii_uppercase());
assert!(!s3.is_ascii_uppercase());

pub const fn is_ascii_alphabetic_lowercase(&self) -> bool

Checks if the value is composed of ASCII alphabetic lower case characters:

U+0061 ‘a’ ..= U+007A ‘z’,

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "Test".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "Te3t".parse().expect("Failed to parse.");
let s3: TinyAsciiStr<4> = "teSt".parse().expect("Failed to parse.");
let s4: TinyAsciiStr<4> = "test".parse().expect("Failed to parse.");
let s5: TinyAsciiStr<4> = "001z".parse().expect("Failed to parse.");

assert!(!s1.is_ascii_alphabetic_lowercase());
assert!(!s2.is_ascii_alphabetic_lowercase());
assert!(!s3.is_ascii_alphabetic_lowercase());
assert!(s4.is_ascii_alphabetic_lowercase());
assert!(!s5.is_ascii_alphabetic_lowercase());

pub const fn is_ascii_alphabetic_titlecase(&self) -> bool

Checks if the value is composed of ASCII alphabetic, with the first character being ASCII uppercase, and all others ASCII lowercase.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "Test".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "Te3t".parse().expect("Failed to parse.");
let s3: TinyAsciiStr<4> = "teSt".parse().expect("Failed to parse.");
let s4: TinyAsciiStr<4> = "test".parse().expect("Failed to parse.");
let s5: TinyAsciiStr<4> = "001z".parse().expect("Failed to parse.");

assert!(s1.is_ascii_alphabetic_titlecase());
assert!(!s2.is_ascii_alphabetic_titlecase());
assert!(!s3.is_ascii_alphabetic_titlecase());
assert!(!s4.is_ascii_alphabetic_titlecase());
assert!(!s5.is_ascii_alphabetic_titlecase());

pub const fn is_ascii_alphabetic_uppercase(&self) -> bool

Checks if the value is composed of ASCII alphabetic upper case characters:

U+0041 ‘A’ ..= U+005A ‘Z’,

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "Test".parse().expect("Failed to parse.");
let s2: TinyAsciiStr<4> = "Te3t".parse().expect("Failed to parse.");
let s3: TinyAsciiStr<4> = "teSt".parse().expect("Failed to parse.");
let s4: TinyAsciiStr<4> = "TEST".parse().expect("Failed to parse.");
let s5: TinyAsciiStr<4> = "001z".parse().expect("Failed to parse.");

assert!(!s1.is_ascii_alphabetic_uppercase());
assert!(!s2.is_ascii_alphabetic_uppercase());
assert!(!s3.is_ascii_alphabetic_uppercase());
assert!(s4.is_ascii_alphabetic_uppercase());
assert!(!s5.is_ascii_alphabetic_uppercase());

impl<const N: usize> TinyAsciiStr<N>

pub const fn to_ascii_lowercase(self) -> Self

Converts this type to its ASCII lower case equivalent in-place.

ASCII letters ‘A’ to ‘Z’ are mapped to ‘a’ to ‘z’, other characters are unchanged.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "TeS3".parse().expect("Failed to parse.");

assert_eq!(&*s1.to_ascii_lowercase(), "tes3");

pub const fn to_ascii_titlecase(self) -> Self

Converts this type to its ASCII title case equivalent in-place.

The first character is converted to ASCII uppercase; the remaining characters are converted to ASCII lowercase.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "teSt".parse().expect("Failed to parse.");

assert_eq!(&*s1.to_ascii_titlecase(), "Test");

pub const fn to_ascii_uppercase(self) -> Self

Converts this type to its ASCII upper case equivalent in-place.

ASCII letters ‘a’ to ‘z’ are mapped to ‘A’ to ‘Z’, other characters are unchanged.

Examples

use tinystr::TinyAsciiStr;

let s1: TinyAsciiStr<4> = "Tes3".parse().expect("Failed to parse.");

assert_eq!(&*s1.to_ascii_uppercase(), "TES3");

impl<const N: usize> TinyAsciiStr<N>

pub const fn to_unvalidated(self) -> UnvalidatedTinyAsciiStr<N>