Struct tinystr::TinyStr16

pub struct TinyStr16(_);

A tiny string that is from 1 to 16 non-NUL ASCII characters.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "Metamorphosis".parse()
    .expect("Failed to parse.");

assert_eq!(s1, "Metamorphosis");
assert!(s1.is_ascii_alphabetic());

Methods

impl TinyStr16

pub fn from_bytes(bytes: &[u8]) -> Result<Self, Error>

Creates a TinyStr16 from a byte slice.

Examples

use tinystr::TinyStr16;

let s1 = TinyStr16::from_bytes("Testing".as_bytes())
    .expect("Failed to parse.");

assert_eq!(s1, "Testing");

pub const unsafe fn new_unchecked(text: u128) -> Self

An unsafe constructor intended for cases where the consumer guarantees that the input is a little endian integer which is a correct representation of a TinyStr16 string.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "Metamorphosis".parse()
    .expect("Failed to parse.");

let num: u128 = s1.into();

let s2 = unsafe { TinyStr16::new_unchecked(num) };

assert_eq!(s1, s2);
assert_eq!(s2.as_str(), "Metamorphosis");

pub fn as_str(&self) -> &str

Extracts a string slice containing the entire TinyStr16.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "Metamorphosis".parse()
    .expect("Failed to parse.");

assert_eq!(s1.as_str(), "Metamorphosis");

pub fn is_ascii_alphabetic(self) -> bool

Checks if the value is composed of ASCII alphabetic characters:

U+0041 'A' ..= U+005A 'Z', or

U+0061 'a' ..= U+007A 'z'.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "Metamorphosis".parse()
    .expect("Failed to parse.");
let s2: TinyStr16 = "Met3mo4pho!is".parse()
    .expect("Failed to parse.");

assert!(s1.is_ascii_alphabetic());
assert!(!s2.is_ascii_alphabetic());

pub fn is_ascii_alphanumeric(self) -> bool

Checks if the value is composed of ASCII alphanumeric characters:

U+0041 'A' ..= U+005A 'Z', or

U+0061 'a' ..= U+007A 'z', or

U+0030 '0' ..= U+0039 '9'.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "A15bingA1".parse()
    .expect("Failed to parse.");
let s2: TinyStr16 = "[3@w00Fs1".parse()
    .expect("Failed to parse.");

assert!(s1.is_ascii_alphanumeric());
assert!(!s2.is_ascii_alphanumeric());

pub fn is_ascii_numeric(self) -> bool

Checks if the value is composed of ASCII decimal digits:

U+0030 '0' ..= U+0039 '9'.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "31212314141".parse()
    .expect("Failed to parse.");
let s2: TinyStr16 = "3d3d3d3d".parse()
    .expect("Failed to parse.");

assert!(s1.is_ascii_numeric());
assert!(!s2.is_ascii_numeric());

pub fn to_ascii_lowercase(self) -> Self

Converts this type to its ASCII lower case equivalent in-place.

ASCII letters 'A' to 'Z' are mapped to 'a' to 'z', other characters are unchanged.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "MeTAmOrpHo3sis".parse()
    .expect("Failed to parse.");

assert_eq!(s1.to_ascii_lowercase(), "metamorpho3sis");

pub fn to_ascii_titlecase(self) -> Self

Converts this type to its ASCII title case equivalent in-place.

First character, if is an ASCII letter 'a' to 'z' is mapped to 'A' to 'Z', other characters are unchanged.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "metamorphosis".parse()
    .expect("Failed to parse.");

assert_eq!(s1.to_ascii_titlecase(), "Metamorphosis");

pub fn to_ascii_uppercase(self) -> Self

Converts this type to its ASCII upper case equivalent in-place.

ASCII letters 'a' to 'z' are mapped to 'A' to 'Z', other characters are unchanged.

Examples

use tinystr::TinyStr16;

let s1: TinyStr16 = "Met3amorphosis".parse()
    .expect("Failed to parse.");

assert_eq!(s1.to_ascii_uppercase(), "MET3AMORPHOSIS");