tinystr

Module int_ops

Structs

Aligned4

Internal helper struct that performs operations on aligned integers. Supports strings up to 4 bytes long.

Aligned8

Internal helper struct that performs operations on aligned integers. Supports strings up to 8 bytes long.